requirements.txt
torch==2.2.2
torchvision==0.17.2
xformers==0.0.25
pytorch_lightning==1.4.2
einops
open-clip-torch==2.23.0
omegaconf
torchmetrics==0.6.0
triton==2.2.0
opencv-python-headless
scipy
matplotlib
lpips
gradio
chardet
transformers
facexlib
numpy==1.26.1
pyiqa
openai_clip
imgaug
imageio
scikit-image
lazy_loader
